+++ src/full/Agda/TypeChecking/Pretty/Warning.hs
@@ -627,6 +627,11 @@ prettyWarning = \case
       [pretty e] ++ pwords "cannot appear by itself. It needs to be the argument to" ++
       pwords "a function expecting an instance argument."
 
+    IllegalDeclarationInDataDefinition ds -> vcat
+      [ "Illegal declaration in data type definition"
+      , nest 2 $ vcat $ fmap pretty ds
+      ]
+
     InvalidDisplayForm x reason -> fsep $ concat
         [ pwords "Ignoring invalid display form for"
         , [ prettyTCM x ]